% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badaj_gre.R
\name{badaj_gre}
\alias{badaj_gre}
\title{Badanie gry SuperFarmer}
\usage{
badaj_gre(namestrategy, rounds, fox_leaves_rabbit = 0, wolf_eats_dog = 0,
  die1 = dice1, die2 = dice2)
}
\arguments{
\item{namestrategy}{Nazwa wykonywanej strategii.}

\item{rounds}{Liczba wykonanych powtórzeń danej gry.}

\item{fox_leaves_rabbit}{Parametr mówiący czy lis zostawia jednego królika. Domyślna wartość 0 (nie zostawia).}

\item{wolf_eats_dog}{Parametr mówiący czy wilk zjada małego psa. Domyślna wartość 0 (nie zostawia).}

\item{die1}{Kostka nr1.}

\item{die2}{Kostka nr2.}
}
\value{
Lista trzyelementowa. Pierwszym elementem jest wektor długości kolejnych
 gier, drugim - podstawowe statystyki rozkładu czasu gry (wynik funkcji summary),
 trzecim elementem jest histogram przedstawiający rozkład otrzymanych rezultatów.
}
\description{
Funkcja \code{badaj_gre} bada rozkład czasu gry SuperFarmer.
}
\examples{
badaj_gre(strategia_PDMS, 1000, 0, 0, dice1, dice2)
badaj_gre(strategia_PDMS, 1000)

}

