% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funkcje_pomocnicze.R
\name{clear.all.counts}
\alias{clear.all.counts}
\title{Zmiana liczebności całego stada na zerowy}
\usage{
clear.all.counts(game)
}
\arguments{
\item{game}{Tabela gry SuperFarmer.}
}
\description{
Funkcja \code{clear.all.counts}  zeruje liczbę wszystkich zwierząt.
}
\examples{
clear.all.counts(game)

}

