% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategia_DKA.R
\name{czy_stac_nas}
\alias{czy_stac_nas}
\title{Czy stać nas}
\usage{
czy_stac_nas(nr_zwierzecia, tmp_zwierzeta)
}
\arguments{
\item{nr_zwierzecia}{numer odpowiadający zwierzęciu, dla którego chcemy
sprawdzić czy możliwa jest wymiana.}

\item{tmp_zwierzeta}{tabela zawierająca liczby posiadanych zwierząt,
liczby zwierząt pozostałych w stadzie oraz wartości zwierząt w przeliczeniu
 na króliki (doładna postać tabeli w przykładzie).}
}
\value{
TRUE jeśli możliwa jest wymiana, FALSE w przeciwnym przypadku.
}
\description{
Funkcja \code{czy_stac_nas} sprawdza, czy jest wystarczająco dużo tańszych
zwierząt, by wymienić je na dane zwierzę.
}
\examples{
zwierzeta <- data.frame(
krolik=c(1, 1, 59),
owca=c(0, 6, 24),
swinia=c(0, 12, 20),
krowa=c(1, 36, 11),
duzy_pies=c(1, 36, 1),
kon=c(2, 72, 4),
maly_pies=c(0,6,4),
row.names=c("w_zagrodzie", "krolikowartosc", "w_stadzie"))
zwierzeta
czy_stac_nas(6, zwierzeta)

}

