% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dice1.R
\docType{data}
\name{dice1}
\alias{dice1}
\title{Kość 1}
\format{12 elementowy wektor zawierający nazwy zwierząt}
\description{
Wektor spośród którego będziemy losować zwierzęta. Zawiera on nazwy zwierząt:
\itemize{
\item rabbit x 6
\item sheep x 3
\item pig
\item cow
\item wolf
}
}
\examples{
library("SuperFarmerDA")
dice1
}
\keyword{datasets}

