% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dice2.R
\docType{data}
\name{dice2}
\alias{dice2}
\title{Kość 2}
\format{12 elementowy wektor zawierający nazwy zwierząt na kostkach.}
\description{
Wektor spośród którego będziemy losować zwierzęta w naszej grze.
Zawiera on nazwy zwierząt:
\itemize{
\item rabbit x 6
\item sheep x 2
\item pig x 2
\item cow
\item fox
}
}
\examples{
library("SuperFarmerDA")
dice2
}
\keyword{datasets}

