% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategia_DKA.R
\name{dodaj_do_zagrody}
\alias{dodaj_do_zagrody}
\title{Dodawanie zwierząt do zagrody}
\usage{
dodaj_do_zagrody(nr_zwierzecia, liczba, tmp_zwierzeta)
}
\arguments{
\item{nr_zwierzecia}{numer odpowiadający zwierzęciu, które chcemy dodać do zagrody}

\item{liczba}{ile zwierząt ma być dodanych}

\item{tmp_zwierzeta}{tabela zawierająca liczby posiadanych zwierząt,
liczby zwierząt pozostałych w stadzie oraz wartości zwierząt w przeliczeniu
 na króliki (doładna postać tabeli w przykładzie).}
}
\value{
Tabela z uaktualnionymi wartościami posiadanych zwierząt.
}
\description{
Funkcja ta dodaje określoną liczbę danego zwierzęcia do aktualnego stanu naszej
 zagrody i odejmuje tę liczbę od stanu stada.
}
\examples{
zwierzeta <- data.frame(
krolik=c(1, 1, 59),
owca=c(0, 6, 24),
swinia=c(0, 12, 20),
krowa=c(1, 36, 11),
duzy_pies=c(1, 36, 1),
kon=c(2, 72, 4),
maly_pies=c(0,6,4),
row.names=c("w_zagrodzie", "krolikowartosc", "w_stadzie"))
zwierzeta
dodaj_do_zagrody(2, 1, zwierzeta)

}

