% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badaj_gre.R
\name{examine.game}
\alias{examine.game}
\title{Badanie gry SuperFarmer}
\usage{
examine.game(namestrategy, rounds, fox_leaves_rabbit = 0, wolf_eats_dog = 0,
  die1 = dice1, die2 = dice2)
}
\arguments{
\item{namestrategy}{Nazwa wykonywanej strategii.}

\item{rounds}{Liczba wykonanych powtórzeń danej gry.}

\item{fox_leaves_rabbit}{Parametr mówiący czy lis zostawia jednego królika. Domyślna wartość 0 (nie zostawia).}

\item{wolf_eats_dog}{Parametr mówiący czy wilk zjada małego psa. Domyślna wartość 0 (nie zostawia).}

\item{die1}{Kostka nr1.}

\item{die2}{Kostka nr2.}
}
\value{
Lista dwuelementowa. Pierwszym elementem jest wynik funkcji summary od wektora
długości kolejnych gier (podstawowe statystyki rozkładu), drugim - wykres rozkładu czasu gry.
}
\description{
Funkcja \code{examine.game} bada rozkład czasu gry SuperFarmer.
}
\examples{
examine.game(strategy_PDMS,100,0,0, dice1, dice2)

}

