% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategia_PDMS.R
\name{exchange.horse}
\alias{exchange.horse}
\title{Wymień konia}
\usage{
exchange.horse(game)
}
\arguments{
\item{game}{Tabela game zawierająca informację dotyczącego posiadanego stada w chwili obecnej.}
}
\value{
Tablica game po sprawdzeniu warunku wymiany drugiego konia.
}
\description{
Funkcja \code{exchange.horse} zamienia zwięrzęta na pozostałe zwierzęta.
Wymiana dokonuje się  w sytuacji posiadania więcej niż jednego konia.
}
\examples{
exchange.horse(game)

}

