% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funkcje_pomocnicze.R
\name{exchange.two.animals}
\alias{exchange.two.animals}
\title{Wymiana dwojga zwierząt}
\usage{
exchange.two.animals(game, animal1, animal2, animal1count, animal2count)
}
\arguments{
\item{game}{Tabela gry SuperFarmer.}

\item{animal1}{Zwierzę które chcemy zamienić.}

\item{animal2}{Zwierzę które chcemy dostać.}

\item{animal1count}{Liczba zwierząt które zamieniamy.}

\item{animal2count}{Liczba zwierząt które powinniśmy dostać jeśli nie przekroczyliśmy limitu.}
}
\description{
Funkcja \code{exchange.two.animals} zwierzę nr 1 na zwierzę nr 2.
}
\examples{
exchange.two.animals(game,"pig","rabbit",1,12)
exchange.two.animals(game,"pig","rabbit",1,get.value(game,"pig"))

}

