% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gra.R
\docType{data}
\name{game}
\alias{game}
\title{Tabela gry SuperFarmer}
\format{macierz z 7-ma wierszami i 4-ma kolumnami}
\description{
Zbiór danych na którym będą wykonywane strategie gry SuperFarmer.
Informacje dotyczące zwierząt są zawarte w wierszach tabeli.
}
\details{
Przyjęta kolejność zwierząt w tabeli game:
\itemize{
\item rabbit
\item small_dog
\item sheep
\item pig
\item big_dog
\item cow
\item horse
}

Poszczególne kolumny zawierają następujące informacje:
\itemize{
\item value. Wartość zwierzęcia liczona w królikach.
\item max.count. Maksymalnie dopuszczalna liczba zwierząt danego gatunku.
\item win.condition. Minimalny stan zwierząt pozwalający ukończyć grę.
\item count. Liczba obecnie posiadanych zwierząt.
}
}
\examples{
library("SuperFarmerDA")
game
}
\keyword{datasets}

