% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategia_duzokrolikow.R
\name{get.all}
\alias{get.all}
\title{Kupowanie ostatniego zwierzęcia potrzebnego do wygranej.}
\usage{
get.all(game)
}
\arguments{
\item{game}{Tabela gry.}
}
\description{
Funkcja \code{get.all} kupuje krowę jeśli jest mamy wszystkie inne zwierzęta.
}
\examples{
get.small.dog(game,41,1)

}

