% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategia_duzokrolikow.R
\name{get.extra.animal}
\alias{get.extra.animal}
\title{Dodanie dodatkowego zwierzęcia}
\usage{
get.extra.animal(game, rabbitcount, animal, animalcount)
}
\arguments{
\item{game}{Tabela gry.}

\item{rabbitcount}{Liczba królików przy których dokonujemy wymiany.}

\item{animal}{Nazwa zwierzęcia, które chcemy dostać.}

\item{animalcount}{Liczba zwierząt, które chcemy dostać.}
}
\description{
Funkcja \code{get.extra.animal} zamiana króliki na zwierzęta.
Zamienia ona króliki na wskazane przez nas zwierzę w sytuacji gdy mamy więcej królików od zadelkarowanej wartości.
}
\examples{
get.extra.animal(game,45,"sheep",2)

}

