% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategia_PDMS.R
\name{get.horse}
\alias{get.horse}
\title{Zdobądź konia}
\usage{
get.horse(game)
}
\arguments{
\item{game}{Tabela game zawierająca informację dotyczącego posiadanego stada w chwili obecnej.}
}
\value{
Tabela game po sprawdzeniu warunku otrzymania konia.
}
\description{
Funkcja \code{get.horse} sprawdza czy możemy już otrzymać konia.
}
\examples{
get.horse(game)
get.horse(game)

}

