% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategia_duzokrolikow.R
\name{get.initial.rabbits}
\alias{get.initial.rabbits}
\title{Króliki startowe}
\usage{
get.initial.rabbits(game, rabbits)
}
\arguments{
\item{game}{Tabela gry.}

\item{rabbits}{Liczba królików, którą chcemy zachować.}
}
\description{
Funkcja \code{get.initial.rabbits} ma na celu wymianę zwierząt na króliki dopóki
 nie osiągniemy oczekiwanej przez nas liczby.
}
\examples{
get.initial.rabbits(game,12)

}

