% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategia_duzokrolikow.R
\name{get.more.rabbits}
\alias{get.more.rabbits}
\title{Wymiana droższych zwierząt na króliki}
\usage{
get.more.rabbits(game, maxrabbit, farm = c("sheep", "pig"))
}
\arguments{
\item{game}{Tabela gry.}

\item{maxrabbit}{Liczba królików, której nie chcemy przekroczyć. Wymiana,
która dawałaby większą liczbę króliów nie wykona się.}

\item{farm}{wektor zawierający nazwy zwierząt które chcialibyśmy wymienić na króliki.}
}
\description{
Funkcja \code{get.more.rabbits} wymienia (domyślnie) świnię lub owcę na króliki.
}
\examples{
get.more.rabbits(game, 40)

}

