% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategia_duzokrolikow.R
\name{get.small.dog}
\alias{get.small.dog}
\title{Branie małych psów}
\usage{
get.small.dog(game, count1, count2)
}
\arguments{
\item{game}{Tabela gry.}

\item{count1}{Minimalna liczba królików przy których dokonujemy wymiany na małego psa}

\item{count2}{Liczba obecnie posiadanych małych psów}
}
\description{
Funkcja \code{get.small.dog} wymienia króliki na małego psa.
}
\examples{
get.small.dog(game,41,1)

}

