% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funkcje_symulujace_gre.R
\name{gra}
\alias{gra}
\title{Jednokrotne wykonanie symulacji  gry SuperFarmer}
\usage{
gra(strategyname, fox_leaves_rabbit = 0, wolf_eats_dog = 0,
  die1 = SuperFarmerDA::dice1, die2 = SuperFarmerDA::dice2)
}
\arguments{
\item{strategyname}{Nazwa wykonywanej strategii.}

\item{fox_leaves_rabbit}{Parametr mówiący czy lis zostawia jednego królika. Domyślna wartość 0 (nie zostawia).}

\item{wolf_eats_dog}{Parametr mówiący czy wilk zjada małego psa. Domyślna wartość 0 (nie zostawia).}

\item{die1}{Kostka nr1.}

\item{die2}{Kostka nr2.}
}
\description{
Funkcja \code{gra} symuluje całą grę jeden raz.
Jako wartość zwraca liczbę rund potrzebnych do ukończenia gry.
}
\examples{
gra(strategia_PDMS,0,1,dice1,dice2)

}

