% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funkcje_symulujace_gre.R
\name{make.move}
\alias{make.move}
\title{Pojedyńczy ruch w grze SuperFarmer}
\usage{
make.move(game, strategy, fox_leaves_rabbit = 0, wolf_eats_dog = 0,
  die1 = dice1, die2 = dice2)
}
\arguments{
\item{game}{Tabela game zawierająca informację dotyczącego posiadanego stada w chwili obecnej.}

\item{strategy}{Nazwa wykonywanej strategii.}

\item{fox_leaves_rabbit}{Parametr mówiący czy lis zostawia jednego królika. Domyślna wartość 0 (nie zostawia).}

\item{wolf_eats_dog}{Parametr mówiący czy wilk zjada małego psa. Domyślna wartość 0 (nie zostawia).}

\item{die1}{Kostka nr1.}

\item{die2}{Kostka nr2.}
}
\description{
Funkcja \code{make.move} wykonuje jeden ruch w grze SuperFarmer.
Jako wartość zwraca tabelę game po wykonaniu jednego ruchu w grze.
}
\examples{
make.move(game,strategia_DKA,0,0,dice1,dice2)

}

