% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funkcje_fazy_rzutu_kostka.R
\name{multiply.animals}
\alias{multiply.animals}
\title{Dodaj po wylosowaniu}
\usage{
multiply.animals(game, animal1, animal2)
}
\arguments{
\item{game}{Tabela game zawierająca informację dotyczącego posiadanego stada w chwili obecnej.}

\item{animal1}{Zwierzę otrzymane w wyniku losowania na kostce nr 1.}

\item{animal2}{Zwierzę otrzymane w wyniku losowania na kostce nr 2.}
}
\value{
Tablica game po wykonaniu rzutu kostką.
}
\description{
Funkcja \code{multiply.animals} dodaje zwierzęta do stada po wykonaniu rzutu kostką.
}
\examples{
multiply.animals(game,"rabbit","sheep")
multiply.animals(game,"pig","pig")
multiply.animals(game,"cow","pig")

}

