% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategia_DKA.R
\name{oddaj_do_stada}
\alias{oddaj_do_stada}
\title{Oddawanie zwierząt do stada}
\usage{
oddaj_do_stada(nr_zwierzecia, liczba, tmp_zwierzeta)
}
\arguments{
\item{nr_zwierzecia}{numer odpowiadający zwierzęciu, które chcemy oddać do stada}

\item{liczba}{ile zwierząt ma być oddanych}

\item{tmp_zwierzeta}{tabela zawierająca liczby posiadanych zwierząt,
liczby zwierząt pozostałych w stadzie oraz wartości zwierząt w przeliczeniu
 na króliki (doładna postać tabeli w przykładzie).}
}
\value{
Tabela z uaktualnionymi wartościami posiadanych zwierząt.
}
\description{
Funkcja ta odejmuje określoną liczbę danego zwierzęcia z aktualnego stanu naszej
 zagrody i dodaje tę liczbę do stanu stada.
}
\examples{
zwierzeta <- data.frame(
krolik=c(1, 1, 59),
owca=c(0, 6, 24),
swinia=c(0, 12, 20),
krowa=c(1, 36, 11),
duzy_pies=c(1, 36, 1),
kon=c(2, 72, 4),
maly_pies=c(0,6,4),
row.names=c("w_zagrodzie", "krolikowartosc", "w_stadzie"))
zwierzeta
oddaj_do_stada(1, 1, zwierzeta)

}

