% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funkcje_symulujace_gre.R
\name{play}
\alias{play}
\title{Jednokrotne wykonanie symulacji  gry SuperFarmer}
\usage{
play(strategyname, fox_leaves_rabbit = 0, wolf_eats_dog = 0, die1, die2)
}
\arguments{
\item{strategyname}{Nazwa wykonywanej strategii.}

\item{fox_leaves_rabbit}{Parametr mówiący czy lis zostawia jednego królika. Domyślna wartość 0 (nie zostawia).}

\item{wolf_eats_dog}{Parametr mówiący czy wilk zjada małego psa. Domyślna wartość 0 (nie zostawia).}

\item{die1}{Kostka nr1.}

\item{die2}{Kostka nr2.}
}
\description{
Funkcja \code{play} symuluje całą grę jeden raz.
Jako wartość zwraca liczbę rund potrzebnych do ukończenia gry.
}
\examples{
play(strategy_PDMS,0,1,die1,die2)

}

