% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategia_duzokrolikow.R
\name{strategia_maxrabbit}
\alias{strategia_maxrabbit}
\title{Strategia maxrabbit}
\usage{
strategia_maxrabbit(vector)
}
\arguments{
\item{vector}{Wektor z liczebnością stada.}
}
\value{
Wektor z zadaną liczebnością stada po wykonaniu strategii.
}
\description{
Strategia mająca na celu otrzymanie jak największej liczby królików.
}
\details{
Strategia jest strategią wiele na wiele. Składa się z następujących etapów:
\itemize{
 \item Konwersji wektora do tabeli gra - możliwość uruchomienia  strategii zewnętrznych w naszym pakiecie.
 \item Zdobycie oczekiwanej przez nas liczby królików.
 \item Zdobycie małego psa.
 \item Zdobycie przynajmniej 40 królików.
 \item Zdobycie drugiego małego psa.
 \item Zamiany zwierząt na dodatkowe w przypadku gdy mamy więcej niż 40 królików.
 \item Kupna konia.
 \item Zamiany drugiego konia na pozostałe zwierzęta.
}
}
\examples{
strategia_maxrabbit(c(8,1,3,0,0,0,1))

}
\author{
Przemysław Dycha, Dominik Ambroziak
}

