% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategia_PDMS.R
\name{strategia_PDMS}
\alias{strategia_PDMS}
\title{Strategia PDMS}
\usage{
strategia_PDMS(vector)
}
\arguments{
\item{vector}{Wektor z liczebnością stada.}
}
\value{
Wektor z zadaną liczebnością stada po wykonaniu strategii.
}
\description{
Strategia z pierwszej fazy używana przez grupę złożoną z Przemysława Dychy i Mateusza Siwca.
Funkcja \code{strategia_PDMS}
}
\details{
Strategia jest strategią wiele na wiele. Składa się z następujących etapów:
\itemize{
 \item Konwersji wektora do tabeli gra - możliwość uruchomienia  strategii zewnętrznych w naszym pakiecie.
 \item Zamiany nadmiarowych zwierząt na króliki.
 \item Wymiany nadmiarowych królików na brakujące zwierzęta.
 \item Wymiany zwierząt na konia.
 \item Wymiany nadmiarowego konia na zwierzęta - zakończenie gry.
 \item Konwersji do wektora - możliwość uruchomienia w innych pakietach.
}
}
\examples{
strategia_PDMS(c(8,1,3,0,0,0,1))

}
\author{
Przemysław Dycha, Mateusz Siwiec
}

