% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategia2.R
\name{strategy_test}
\alias{strategy_test}
\title{Strategia testowa}
\usage{
strategy_test(vector)
}
\arguments{
\item{vector}{Wektor zawierający informację dotyczącego posiadanego stada w chwili obecnej.}
}
\description{
Funkcja \code{strategy_test}
}
\details{
Strategia jest strategią jeden  na wiele. Można wymieniać króliki tylko na zwierzęta o wyższej wartości.
\itemize{
 \item Konwersji wektora do tabeli gra - możliwość uruchomienia  strategii zewnętrznych w naszym pakiecie.
 \item Wymiany nadmiarowych królików na brakujące zwierzęta.
 \item Wymiany zwierząt na konia.
 \item Wymiany nadmiarowego konia na zwierzęta - zakończenie gry.
 \item Konwersji do wektora - możliwość uruchomienia w innych pakietach.
}
}
\examples{
strategy_test(c(8,1,3,0,0,0,1))

}

