% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategia_duzokrolikow.R
\name{value.in.rabbit}
\alias{value.in.rabbit}
\title{Wartość stada w królikach}
\usage{
value.in.rabbit(game)
}
\arguments{
\item{game}{Tabela gry.}
}
\description{
Funkcja \code{value.in.rabbit} zwraca informację o wartości posiadanego statada.
}
\examples{
value.in.rabbit(game)

}

