% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategia_duzokrolikow.R
\name{value.in.rabbit.without.animal}
\alias{value.in.rabbit.without.animal}
\title{Wartość stada bez wskazanych zwierząt}
\usage{
value.in.rabbit.without.animal(game, animal)
}
\arguments{
\item{game}{Tabela gry.}

\item{animal}{Nazwa zwierzęcia.}
}
\description{
Funkcja \code{value.in.rabbit.without.animal} zwraca informację o wartości posiadanego statada bez wskazanego zwierzęcia.
}
\examples{
value.in.rabbit.without.animal(game,"rabbit")

}

