% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wizytowka_ADPD.R
\name{wizytowka_ADPD}
\alias{wizytowka_ADPD}
\title{Wizytówka}
\usage{
wizytowka_ADPD(strategia, N = 10000)
}
\arguments{
\item{strategia}{Tabela gry SuperFarmer.}

\item{N}{Liczba powtórzeń dla których powinna być wykonana.}
}
\value{
Plik pdf z wynikami strategii.
}
\description{
Funkcja \code{wizytowka_ADPD} tworzy wizytówkę dla strategii.
Funkcja automatycznie wczytuje nazwę strategii zgodną ze wzorcem strategia_
Do uruchomienia pakietu konieczny jest pakiet dplyr
}
\examples{
wizytowka_ADPD(strategia_DKA,10)

}

