% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wizytowka_PA.R
\name{wizytowka_PA}
\alias{wizytowka_PA}
\title{Wizytówka}
\usage{
wizytowka_PA(strategia, nazwa_strategii, ile_powtorzen = 10000)
}
\arguments{
\item{strategia}{- funkcja zawierająca strategię, którą chcemy zbadać}

\item{nazwa_strategii}{- argument tekstowy, nazwa badanej strategii}

\item{ile_powtorzen}{- argument liczbowy mówiący ile razy ma być przeprowadzona symulacja gry}
}
\value{
Ramka obiektów graficznych z wizualnym porównaniem wydajności badanej strategii obok strategii strategy_maxrabbit
}
\description{
Funkcja tworząca wizualne porównanie wydajności badanej strategii ze strategią strategia_maxrabbit.
W otrzymanym pliku pdf znajduje się plansza z nałożonymi na siebie wykresami rozkładów długości gry dla obu strategii,  plansza z wykresami słupkowymi obrazującymi pierwsze i ostatnie decyle tych rozkładów oraz tabela z ich najważniejszymi parametrami.
}
\examples{
wizytowka_PA(SuperFarmerDA::strategia_PDMS, "strategia_PDMS", 500)

}
\author{
Katarzyna Pyrtek, \email{kat.pyrtek@gmail.com}

Dominik Ambroziak, \email{ambroziak.d2@gmail.com}
}

