% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funkcje_fazy_rzutu_kostka.R
\name{wolf.reaction}
\alias{wolf.reaction}
\title{Reakcja na wilka}
\usage{
wolf.reaction(game, die1, takesmalldog = 0)
}
\arguments{
\item{game}{Tabela game zawierająca informację dotyczącego posiadanego stada w chwili obecnej.}

\item{die1}{Zwierzę otrzymane w wyniku losowania na kostce nr 1.}

\item{takesmalldog}{Parametr dodatkowy mówiący czy wilk zjada małego psa.
W przypadku gdy wilk zjada małego psa należy wstawić 1. W przeciwnym przypadku 0.}
}
\value{
Tablica game po wykonaniu rzutu kostką.
}
\description{
Funkcja \code{wolf.reaction} symuluje zmianę stada po wylosowaniu wilka.
}
\examples{
wolf.reaction(game,"rabbit",0)
wolf.reaction(game,"wolf",1)

}

