% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategia_DKA.R
\name{wymiana_na_drozsze}
\alias{wymiana_na_drozsze}
\title{Wymiana na droższe}
\usage{
wymiana_na_drozsze(tmp_zwierzeta)
}
\arguments{
\item{tmp_zwierzeta}{tabela zawierająca liczby posiadanych zwierząt,
liczby zwierząt pozostałych w stadzie oraz wartości zwierząt w przeliczeniu
 na króliki (dokładna postać tabeli w przykładzie).}
}
\value{
Tabela z uaktualnionymi wartościami posiadanych zwierząt.
}
\description{
Funkcja ta zamienia posiadane zwierzęta na jedno, możliwie jak największe, zwierzę.
}
\details{
Rozpoczynając od najdroższego zwierzęcia sprawdzane jest, czy można
je otrzymać wymieniając tańsze zwierzęta. Dokonuje się tylko jedna wymiana. Wymiana nie dokona się, jeśli
 powodowałaby wyzerowanie stanu królików.
}
\examples{
zwierzeta <- data.frame(
krolik=c(7, 1, 53),
owca=c(1, 6, 23),
swinia=c(0, 12, 20),
krowa=c(0, 36, 12),
duzy_pies=c(0, 36, 2),
kon=c(0, 72, 6),
maly_pies=c(0,6,4),
row.names=c("w_zagrodzie", "krolikowartosc", "w_stadzie"))
zwierzeta
wymiana_na_drozsze(zwierzeta)

}

