% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategia_DKA.R
\name{wymiana_na_tansze}
\alias{wymiana_na_tansze}
\title{Wymiana na tańsze}
\usage{
wymiana_na_tansze(tmp_zwierzeta)
}
\arguments{
\item{tmp_zwierzeta}{tabela zawierająca liczby posiadanych zwierząt,
liczby zwierząt pozostałych w stadzie oraz wartości zwierząt w przeliczeniu
 na króliki (dokładna postać tabeli w przykładzie).}
}
\value{
Tabela z uaktualnionymi wartościami posiadanych zwierząt.
}
\description{
Funkcja ta zamienia jedno zwierzę na kilka tańszych zwierząt. Jest wywoływana
w funkcji strategia_DKA tylko wtedy, gdy wartość posiadanych zwierząt jest
wystarczająca do otrzymania wszystkich zwierząt wymaganych do zakończenia gry.
}
\examples{
zwierzeta <- data.frame(
krolik=c(1, 1, 59),
owca=c(0, 6, 24),
swinia=c(0, 12, 20),
krowa=c(1, 36, 11),
duzy_pies=c(1, 36, 1),
kon=c(2, 72, 4),
maly_pies=c(0,6,4),
row.names=c("w_zagrodzie", "krolikowartosc", "w_stadzie"))
zwierzeta
wymiana_na_tansze(zwierzeta)

}

