% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategia_duzokrolikow.R
\name{buy.horse.for.animals}
\alias{buy.horse.for.animals}
\title{Kupno konia}
\usage{
buy.horse.for.animals(game, rabbitcount)
}
\arguments{
\item{game}{Tabela gry.}

\item{rabbitcount}{Liczba królików, które chcemy zostawić.}
}
\description{
Funkcja \code{get.extra.animal} dodaje nam konia do stada.
}
\examples{
buy.horse.for.animals(game,12)

}

