% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategia_duzokrolikow.R
\name{get.small.dog}
\alias{get.small.dog}
\title{Branie małych psów}
\usage{
get.small.dog(game, count1, count2)
}
\arguments{
\item{game}{Tabela gry.}

\item{count1}{Liczba królików przy których dokonujemy wymiany na psa}

\item{count2}{Liczba obecnie posiadanych małych psów}
}
\description{
Funkcja \code{initial.rabbits} ma na celu wymianę zwierząt na króliki dopóki nie osiągniemy oczekiwanej przez nas liczby.
}
\examples{
get.small.dog(game,41,1)

}

